fix(difficulty): Make it a double value instead

+++ api/src/main/java/dev/jsinco/brewery/recipe/Recipe.java
@@ -8,7 +8,7 @@ public interface Recipe<I> {
 
     String getRecipeName();
 
-    int getBrewDifficulty();
+    double getBrewDifficulty();
 
     List<BrewingStep> getSteps();
